refactor(elements): Relocate machines
