+++ packages/elements/src/internals/machines/form/form.context.ts
@@ -1,7 +1,7 @@
 import { createActorContext } from '@xstate/react';
 import type { SnapshotFrom } from 'xstate';
 
-import { FormMachine } from '~/internals/machines/form/form.machine';
+import { FormMachine } from '~/internals/machines/form';
 import { consoleInspector } from '~/internals/utils/inspector';
 
 export type SnapshotState = SnapshotFrom<typeof FormMachine>;
