+++ packages/elements/src/internals/machines/sign-in/index.ts
@@ -1,7 +1,16 @@
 export { SignInFirstFactorMachine, SignInSecondFactorMachine } from './verification.machine';
 export { SignInRouterMachine, SignInRouterMachineId } from './router.machine';
 export { SignInStartMachine, SignInStartMachineId } from './start.machine';
+export { SignInResetPasswordMachine, SignInResetPasswordMachineId } from './reset-password.machine';
+
+export { SignInSafeIdentifierSelector, SignInSalutationSelector } from './router.selectors';
 
 export type { TSignInRouterMachine } from './router.machine';
 export type { TSignInStartMachine } from './start.machine';
 export type { TSignInFirstFactorMachine, TSignInSecondFactorMachine } from './verification.machine';
+export type { TSignInResetPasswordMachine } from './reset-password.machine';
+
+export * from './verification.types';
+export * from './router.types';
+export * from './start.types';
+export * from './reset-password.types';
