+++ packages/elements/src/internals/machines/sign-in/start.machine.ts
@@ -1,29 +1,41 @@
-import type { SignUpResource } from '@clerk/types';
+import type { SignInResource } from '@clerk/types';
 import type { ActorRefFrom } from 'xstate';
 import { fromPromise, not, sendTo, setup } from 'xstate';
 
-import { SIGN_UP_DEFAULT_BASE_PATH } from '~/internals/constants';
-import type { FormFields } from '~/internals/machines/form/form.types';
-import { sendToLoading } from '~/internals/machines/shared.actions';
-import type { TSignUpRouterMachine } from '~/internals/machines/sign-up/machines';
-import type { SignUpStartSchema } from '~/internals/machines/sign-up/types';
-import { fieldsToSignUpParams } from '~/internals/machines/sign-up/utils';
-import { ThirdPartyMachine } from '~/internals/machines/third-party/machine';
+import { SIGN_IN_DEFAULT_BASE_PATH } from '~/internals/constants';
+import type { FormFields } from '~/internals/machines/form';
+import { sendToLoading } from '~/internals/machines/shared';
 import { assertActorEventError } from '~/internals/machines/utils/assert';
 
-export type TSignUpStartMachine = typeof SignUpStartMachine;
+import type { TSignInRouterMachine } from './router.machine';
+import type { SignInStartSchema } from './start.types';
 
-export const SignUpStartMachineId = 'SignUpStart';
+export type TSignInStartMachine = typeof SignInStartMachine;
 
-export const SignUpStartMachine = setup({
+export const SignInStartMachineId = 'SignInStart';
+
+export const SignInStartMachine = setup({
   actors: {
-    attempt: fromPromise<SignUpResource, { parent: ActorRefFrom<TSignUpRouterMachine>; fields: FormFields }>(
+    attempt: fromPromise<SignInResource, { parent: ActorRefFrom<TSignInRouterMachine>; fields: FormFields }>(
       ({ input: { fields, parent } }) => {
-        const params = fieldsToSignUpParams(fields);
-        return parent.getSnapshot().context.clerk.client.signUp.create(params);
+        const clerk = parent.getSnapshot().context.clerk;
+
+        const password = fields.get('password');
+        const identifier = fields.get('identifier');
+
+        const passwordParams = password?.value
+          ? {
+              password: password.value,
+              strategy: 'password',
+            }
+          : {};
+
+        return clerk.client.signIn.create({
+          identifier: (identifier?.value as string) || '',
+          ...passwordParams,
+        });
       },
     ),
-    thirdParty: ThirdPartyMachine,
   },
   actions: {
     setFormErrors: sendTo(
@@ -42,13 +54,13 @@ export const SignUpStartMachine = setup({
   guards: {
     isExampleMode: ({ context }) => Boolean(context.parent.getSnapshot().context.exampleMode),
   },
-  types: {} as SignUpStartSchema,
+  types: {} as SignInStartSchema,
 }).createMachine({
-  id: SignUpStartMachineId,
+  id: SignInStartMachineId,
   context: ({ input }) => ({
-    basePath: input.basePath || SIGN_UP_DEFAULT_BASE_PATH,
-    formRef: input.form,
+    basePath: input.basePath || SIGN_IN_DEFAULT_BASE_PATH,
     parent: input.parent,
+    formRef: input.form,
     loadingStep: 'start',
   }),
   initial: 'Pending',
@@ -68,7 +80,7 @@ export const SignUpStartMachine = setup({
       tags: ['state:attempting', 'state:loading'],
       entry: 'sendToLoading',
       invoke: {
-        id: 'attemptCreate',
+        id: 'attempt',
         src: 'attempt',
         input: ({ context }) => ({
           parent: context.parent,
