--- packages/elements/src/internals/machines/sign-in/types/shared.types.ts
@@ -1,3 +1,0 @@
-import type { SignInStrategy } from '@clerk/types';
-
-export type SignInStrategyName = SignInStrategy | 'oauth' | 'web3';
