+++ packages/elements/src/internals/machines/sign-in/verification.types.ts
@@ -2,8 +2,9 @@ import type { ClerkAPIResponseError } from '@clerk/shared/error';
 import type { SignInFactor } from '@clerk/types';
 import type { ActorRefFrom, ErrorActorEvent } from 'xstate';
 
-import type { FormMachine } from '~/internals/machines/form/form.machine';
-import type { TSignInRouterMachine } from '~/internals/machines/sign-in/machines';
+import type { FormMachine } from '~/internals/machines/form';
+
+import type { TSignInRouterMachine } from './router.machine';
 
 // ---------------------------------- Tags ---------------------------------- //
