+++ packages/elements/src/internals/machines/sign-up/index.ts
@@ -7,3 +7,8 @@ export type { TSignUpContinueMachine } from './continue.machine';
 export type { TSignUpRouterMachine } from './router.machine';
 export type { TSignUpStartMachine } from './start.machine';
 export type { TSignUpVerificationMachine } from './verification.machine';
+
+export type * from './continue.types';
+export type * from './router.types';
+export type * from './start.types';
+export type * from './verification.types';
