--- packages/elements/src/internals/machines/sign-up/types/index.ts
@@ -1,4 +1,0 @@
-export * from './continue.types';
-export * from './router.types';
-export * from './start.types';
-export * from './verification.types';
