+++ packages/elements/src/react/sign-in/root.tsx
@@ -4,8 +4,8 @@ import { createActor } from 'xstate';
 
 import { SIGN_IN_DEFAULT_BASE_PATH, SIGN_UP_DEFAULT_BASE_PATH } from '~/internals/constants';
 import { FormStoreProvider } from '~/internals/machines/form/form.context';
-import { SignInRouterMachine } from '~/internals/machines/sign-in/machines';
-import type { SignInRouterInitEvent } from '~/internals/machines/sign-in/types';
+import type { SignInRouterInitEvent } from '~/internals/machines/sign-in';
+import { SignInRouterMachine } from '~/internals/machines/sign-in';
 import { consoleInspector } from '~/internals/utils/inspector';
 import { Router, useClerkRouter, useNextRouter } from '~/react/router';
 import { SignInRouterCtx } from '~/react/sign-in/context';
