+++ packages/elements/src/react/sign-in/step/step.tsx
@@ -1,9 +1,13 @@
 import { ClerkElementsRuntimeError } from '~/internals/errors';
 
-import { SignInChooseStrategy, type SignInChooseStrategyProps, SignInForgotPassword } from '../choose-strategy';
-import { SignInStart, type SignInStartProps } from '../start';
-import { SignInVerifications, type SignInVerificationsProps } from '../verifications';
-import { SignInResetPassword, type SignInResetPasswordProps } from './reset-password';
+import type { SignInChooseStrategyProps } from '../choose-strategy';
+import { SignInChooseStrategy, SignInForgotPassword } from '../choose-strategy';
+import type { SignInStartProps } from '../start';
+import { SignInStart } from '../start';
+import type { SignInVerificationsProps } from '../verifications';
+import { SignInVerifications } from '../verifications';
+import type { SignInResetPasswordProps } from './reset-password';
+import { SignInResetPassword } from './reset-password';
 
 export const SIGN_IN_STEPS = {
   start: 'start',
