+++ packages/elements/src/react/sign-up/start.tsx
@@ -1,5 +1,5 @@
-import type { TSignUpStartMachine } from '~/internals/machines/sign-up/machines';
-import { SignUpStartMachine } from '~/internals/machines/sign-up/machines';
+import type { TSignUpStartMachine } from '~/internals/machines/sign-up';
+import { SignUpStartMachine } from '~/internals/machines/sign-up';
 import type { FormProps } from '~/react/common/form';
 import { Form } from '~/react/common/form';
 import { useActiveTags } from '~/react/hooks';
